feat: permissões de grupo de users.
feat: implementação de permissões próprias para grupos de usuários.
