+++ django_restful_api/usuarios/urls.py
@@ -3,6 +3,6 @@ from usuarios.views import UserAuthToken, UserDetail, UserList
 
 urlpatterns = [
     path('', UserList.as_view()),
-    path('<str:username>/', UserDetail.as_view()),
+    path('<slug:username>/', UserDetail.as_view()),
     path('auth/', UserAuthToken.as_view()),
 ]
